mark assembly files as compatible with gc toolchain only
Issue #19 shows that building with llvmgo (or gccgo) currently fails with an unclear message.
This is because it tries to compile .s files that are using a gc-specific format.
A complete solution would introduce .s files that do compile with those alternate toolchains,
but at least we'll now have a clearer error (undefined reference to a bunch of functions)